Moving LinkClock into a separate compilation unit
