+++ lang/LangPrimSource/PyrSched.cpp
@@ -49,23 +49,14 @@
 
 #include "SC_Lock.h"
 
+#include "SC_Clock.h"
 
 #include <boost/sync/semaphore.hpp>
 #include <boost/sync/support/std_chrono.hpp>
 
-#include <ableton/Link.hpp>
-
 static const double dInfinity = std::numeric_limits<double>::infinity();
 
 void runAwakeMessage(VMGlobals *g);
-
-// heaps use an integer timestamp to ensure stable heap order
-struct PyrHeap:
-	PyrObjectHdr
-{
-	PyrSlot count;    // stability count
-	PyrSlot slots[0]; // slots
-};
 
 
 
@@ -242,32 +233,11 @@ using monotonic_clock = std::conditional<std::chrono::high_resolution_clock::is_
 
 static std::chrono::high_resolution_clock::time_point hrTimeOfInitialization;
 
-template <typename DurationType>
-inline double DurToFloat(DurationType dur)
-{
-	using namespace std::chrono;
-	seconds secs         = duration_cast<seconds>(dur);
-	nanoseconds nanosecs = dur - secs;
-
-	return secs.count() + 1.0e-9 * nanosecs.count();
-}
-
-
-static std::chrono::microseconds linkTimeOfInitialization;
-inline std::chrono::microseconds hrToLinkTime(double secs){
-	auto time = std::chrono::duration<double>(secs);
-	return std::chrono::duration_cast<std::chrono::microseconds>(time) + linkTimeOfInitialization;
-}
-inline double linkToHrTime(std::chrono::microseconds micros){
-	return DurToFloat(micros - linkTimeOfInitialization);
-}
-
-
 SCLANG_DLLEXPORT_C void schedInit()
 {
 	using namespace std::chrono;
 	hrTimeOfInitialization     = high_resolution_clock::now();
-	linkTimeOfInitialization   = ableton::link::platform::Clock().micros();
+	initLink();
 
 	syncOSCOffsetWithTimeOfDay();
 	gResyncThread = std::thread(resyncThread);
@@ -683,56 +653,6 @@ new clock:
 		reschedules.
 */
 
-class TempoClock
-{
-public:
-	TempoClock(VMGlobals *inVMGlobals, PyrObject* inTempoClockObj,
-				double inTempo, double inBaseBeats, double inBaseSeconds);
-	virtual ~TempoClock() {}
-	void StopReq();
-	void Stop();
-	void StopAndDelete()
-	{
-		Stop();
-		delete this;
-	}
-
-	void* Run();
-
-	void Add(double inBeats, PyrSlot* inTask);
-	void SetTempoAtBeat(double inTempo, double inBeats);
-	void SetTempoAtTime(double inTempo, double inSeconds);
-	void SetAll(double inTempo, double inBeats, double inSeconds);
-	double ElapsedBeats();
-	void Clear();
-	//void Flush();
-	double GetTempo() const { return mTempo; }
-	double GetBeatDur() const { return mBeatDur; }
-	virtual double BeatsToSecs(double beats) const
-		{ return (beats - mBaseBeats) * mBeatDur + mBaseSeconds; }
-	virtual double SecsToBeats(double secs) const
-		{ return (secs - mBaseSeconds) * mTempo + mBaseBeats; }
-	void Dump();
-
-
-//protected:
-	VMGlobals* g;
-	PyrObject* mTempoClockObj;
-	PyrHeap* mQueue;
-
-	double mTempo; // beats per second
-	double mBeatDur; // 1/tempo
-	double mBeats; // beats
-	double mBaseSeconds;
-	double mBaseBeats;
-	bool mRun;
-	SC_Thread mThread;
-	condition_variable_any mCondition;
-	TempoClock *mPrev, *mNext;
-
-	static TempoClock *sAll;
-
-};
 
 TempoClock *TempoClock::sAll = 0;
 
@@ -987,161 +907,6 @@ void TempoClock::Dump()
 }
 
 
-
-PyrSymbol* s_LinkTempoChanged;
-class LinkClock : public TempoClock
-{
-public:
-	LinkClock(VMGlobals *inVMGlobals, PyrObject* inTempoClockObj,
-				double inTempo, double inBaseBeats, double inBaseSeconds);
-
-	~LinkClock() {}
-
-	void SetTempoAtBeat(double inTempo, double inBeats);
-	void SetTempoAtTime(double inTempo, double inSeconds);
-	void SetAll(double inTempo, double inBeats, double inSeconds);
-	double BeatsToSecs(double beats) const
-	{
-		auto timeline = mLink.captureAppTimeline();
-		double secs = linkToHrTime(timeline.timeAtBeat(beats, mQuantum));
-		return secs;
-	}
-	double SecsToBeats(double secs) const
-	{
-		auto timeline = mLink.captureAppTimeline();
-		double beats = timeline.beatAtTime(hrToLinkTime(secs), mQuantum);
-		return beats;
-	}
-
-	void SetQuantum(double quantum);
-	std::size_t NumPeers() const { return mLink.numPeers(); }
-
-
-protected:
-	ableton::Link mLink;
-	double mQuantum;
-};
-
-
-LinkClock::LinkClock(VMGlobals *inVMGlobals, PyrObject* inTempoClockObj,
-							double inTempo, double inBaseBeats, double inBaseSeconds)
-	: TempoClock(inVMGlobals, inTempoClockObj, inTempo, inBaseBeats, inBaseSeconds),
-	mLink(inTempo * 60.)
-{
-	//quantum
-	int err = slotDoubleVal(&inTempoClockObj->slots[2], &mQuantum);
-	if(err) mQuantum=4.;
-
-	mLink.enable(true);
-	mLink.setTempoCallback([this](double bpm) {
-		double secs = elapsedTime();
-		double tempo = bpm / 60.;
-
-		auto timeline = mLink.captureAppTimeline();
-		double beats = timeline.beatAtTime(hrToLinkTime(secs), mQuantum);
-		mLink.commitAppTimeline(timeline);
-
-		mTempo = tempo;
-		mBeatDur = 1. / tempo;
-		mCondition.notify_one();
-
-		//call sclang callback
-		gLangMutex.lock();
-		g->canCallOS = false;
-		++g->sp;
-		SetObject(g->sp, mTempoClockObj);
-		++g->sp;
-		SetFloat(g->sp, mTempo);
-		++g->sp;
-		SetFloat(g->sp, beats);
-		++g->sp;
-		SetFloat(g->sp, secs);
-		++g->sp;
-		SetObject(g->sp, mTempoClockObj);
-		runInterpreter(g, s_LinkTempoChanged, 5);
-		g->canCallOS = false;
-		gLangMutex.unlock();
-	});
-
-	auto timeline = mLink.captureAppTimeline();
-	auto linkTime = hrToLinkTime(inBaseSeconds);
-	timeline.requestBeatAtTime(inBaseBeats, linkTime, mQuantum);
-	mLink.commitAppTimeline(timeline);
-}
-
-
-void LinkClock::SetAll(double inTempo, double inBeats, double inSeconds)
-{
-
-	auto timeline = mLink.captureAppTimeline();
-	auto linkTime = hrToLinkTime(inSeconds);
-	timeline.setTempo(inTempo * 60., linkTime);
-	timeline.requestBeatAtTime(inBeats,linkTime, mQuantum);
-
-	mTempo = inTempo;
-	mBeatDur = 1. / inTempo;
-
-	mLink.commitAppTimeline(timeline);
-	mCondition.notify_one();
-}
-
-void LinkClock::SetTempoAtBeat(double inTempo, double inBeats)
-{
-	auto timeline = mLink.captureAppTimeline();
-	auto time = timeline.timeAtBeat(inBeats, mQuantum);
-	timeline.setTempo(inTempo*60., time);
-	mLink.commitAppTimeline(timeline);
-}
-
-void LinkClock::SetTempoAtTime(double inTempo, double inSeconds)
-{
-	auto timeline = mLink.captureAppTimeline();
-	timeline.setTempo(inTempo*60., hrToLinkTime(inSeconds));
-	mLink.commitAppTimeline(timeline);
-}
-
-void LinkClock::SetQuantum(double quantum)
-{
-	mQuantum = quantum;
-	mCondition.notify_one();
-}
-
-
-template <typename Clock>
-int prClock_New(struct VMGlobals *g, int numArgsPushed);
-template <typename Clock>
-int prClock_New(struct VMGlobals *g, int numArgsPushed)
-{
-	PyrSlot *a = g->sp - 3;
-	PyrSlot *b = g->sp - 2;
-	PyrSlot *c = g->sp - 1;
-	PyrSlot *d = g->sp;
-
-	double tempo;
-	int err = slotDoubleVal(b, &tempo);
-	if (err) tempo = 1.;
-	if (tempo <= 0.) {
-		error("invalid tempo %g\n", tempo);
-		SetPtr(slotRawObject(a)->slots+1, NULL);
-		return errFailed;
-	}
-
-	double beats;
-	err = slotDoubleVal(c, &beats);
-	if (err) beats = 0.;
-
-	double seconds;
-	err = slotDoubleVal(d, &seconds);
-	if (err) {
-		err = slotDoubleVal(&g->thread->seconds, &seconds);
-		if (err) return err;
-	}
-
-	Clock* clock = new Clock(g, slotRawObject(a), tempo, beats, seconds);
-	SetPtr(slotRawObject(a)->slots+1, clock);
-	return errNone;
-}
-
 int prTempoClock_Free(struct VMGlobals *g, int numArgsPushed);
 int prTempoClock_Free(struct VMGlobals *g, int numArgsPushed)
 {
@@ -1246,32 +1011,6 @@ int prTempoClock_Beats(struct VMGlobals *g, int numArgsPushed)
 	return errNone;
 }
 
-template <typename Clock>
-int prClock_SetBeats(struct VMGlobals *g, int numArgsPushed)
-{
-	PyrSlot *a = g->sp - 1;
-	PyrSlot *b = g->sp;
-
-	Clock *clock = (Clock*)slotRawPtr(&slotRawObject(a)->slots[1]);
-	if (!clock) {
-		error("clock is not running.\n");
-		return errFailed;
-	}
-
-	double beats, seconds;
-	int err;
-
-	err = slotDoubleVal(b, &beats);
-	if (err) return errWrongType;
-
-	err = slotDoubleVal(&g->thread->seconds, &seconds);
-	if (err) return errWrongType;
-
-	clock->SetAll(clock->mTempo, beats, seconds);
-
-	return errNone;
-}
-
 int prTempoClock_Sched(struct VMGlobals *g, int numArgsPushed);
 int prTempoClock_Sched(struct VMGlobals *g, int numArgsPushed)
 {
@@ -1331,103 +1070,6 @@ int prTempoClock_SchedAbs(struct VMGlobals *g, int numArgsPushed)
 	return errNone;
 }
 
-template <typename Clock>
-int prClock_SetTempoAtBeat(struct VMGlobals *g, int numArgsPushed);
-template <typename Clock>
-int prClock_SetTempoAtBeat(struct VMGlobals *g, int numArgsPushed)
-{
-	PyrSlot *a = g->sp - 2;
-	PyrSlot *b = g->sp - 1;
-	PyrSlot *c = g->sp;
-
-	Clock *clock = (Clock*)slotRawPtr(&slotRawObject(a)->slots[1]);
-	if (!clock) {
-		error("clock is not running.\n");
-		return errFailed;
-	}
-	if(clock->mTempo <= 0.f) {
-		error("cannot set tempo from this method. The message 'etempo_' can be used instead.\n");
-		// prTempoClock_SetTempoAtTime can be used.
-		return errFailed;
-	}
-
-	double tempo, beat;
-	int err = slotDoubleVal(b, &tempo);
-	if (err) return errFailed;
-	if (tempo <= 0.) {
-		error("invalid tempo %g. The message 'etempo_' can be used instead.\n", tempo);
-		// prTempoClock_SetTempoAtTime can be used.
-		return errFailed;
-	}
-
-	err = slotDoubleVal(c, &beat);
-	if (err) return errFailed;
-
-	clock->SetTempoAtBeat(tempo, beat);
-
-	return errNone;
-}
-
-template <typename Clock>
-int prClock_SetAll(struct VMGlobals *g, int numArgsPushed);
-template <typename Clock>
-int prClock_SetAll(struct VMGlobals *g, int numArgsPushed)
-{
-	PyrSlot *a = g->sp - 3;
-	PyrSlot *b = g->sp - 2;
-	PyrSlot *c = g->sp - 1;
-	PyrSlot *d = g->sp;
-
-	Clock *clock = (Clock*)slotRawPtr(&slotRawObject(a)->slots[1]);
-	if (!clock) {
-		error("clock is not running.\n");
-		return errFailed;
-	}
-
-	double tempo, beat, secs;
-	int err = slotDoubleVal(b, &tempo);
-	if (err) return errFailed;
-
-	err = slotDoubleVal(c, &beat);
-	if (err) return errFailed;
-
-	err = slotDoubleVal(d, &secs);
-	if (err) return errFailed;
-
-	clock->SetAll(tempo, beat, secs);
-
-	return errNone;
-}
-
-template <typename Clock>
-int prClock_SetTempoAtTime(struct VMGlobals *g, int numArgsPushed);
-template <typename Clock>
-int prClock_SetTempoAtTime(struct VMGlobals *g, int numArgsPushed)
-{
-	PyrSlot *a = g->sp - 2;
-	PyrSlot *b = g->sp - 1;
-	PyrSlot *c = g->sp;
-
-	Clock *clock = (Clock*)slotRawPtr(&slotRawObject(a)->slots[1]);
-	if (!clock) {
-		error("clock is not running.\n");
-		return errFailed;
-	}
-
-	double tempo, sec;
-	int err = slotDoubleVal(b, &tempo);
-	if (err) return errFailed;
-
-	err = slotDoubleVal(c, &sec);
-	if (err) return errFailed;
-
-	clock->SetTempoAtTime(tempo, sec);
-
-	return errNone;
-}
-
-
-
 int prTempoClock_BeatsToSecs(struct VMGlobals *g, int numArgsPushed);
 int prTempoClock_BeatsToSecs(struct VMGlobals *g, int numArgsPushed)
 {
@@ -1469,40 +1111,6 @@ int prTempoClock_SecsToBeats(struct VMGlobals *g, int numArgsPushed)
 
 	return errNone;
 }
-
-
-
-
-int prLinkClock_NumPeers(struct VMGlobals *g, int numArgsPushed);
-int prLinkClock_NumPeers(struct VMGlobals *g, int numArgsPushed)
-{
-	PyrSlot *a = g->sp;
-	LinkClock *clock = (LinkClock*)slotRawPtr(&slotRawObject(a)->slots[1]);
-
-	SetInt(a, clock->NumPeers());
-
-	return errNone;
-}
-
-
-int prLinkClock_SetQuantum(struct VMGlobals *g, int numArgsPushed);
-int prLinkClock_SetQuantum(struct VMGlobals *g, int numArgsPushed)
-{
-	PyrSlot *a = g->sp - 1;
-	PyrSlot *b = g->sp;
-	LinkClock *clock = (LinkClock*)slotRawPtr(&slotRawObject(a)->slots[1]);
-
-	double quantum;
-	int err = slotDoubleVal(b, &quantum);
-	if(err) return errFailed;
-
-	clock->SetQuantum(quantum);
-
-	return errNone;
-}
-
-
-
 
 int prSystemClock_Clear(struct VMGlobals *g, int numArgsPushed);
 int prSystemClock_Clear(struct VMGlobals *g, int numArgsPushed)
@@ -1568,8 +1176,6 @@ void initSchedPrimitives()
 {
 	int base, index=0;
 
-	s_LinkTempoChanged = getsym("prTempoChanged");
-
 	base = nextPrimitiveIndex();
 
 	definePrimitive(base, index++, "_TempoClock_New", prClock_New<TempoClock>, 4, 0);
@@ -1588,14 +1194,6 @@ void initSchedPrimitives()
 	definePrimitive(base, index++, "_TempoClock_SetAll", prClock_SetAll<TempoClock>, 4, 0);
 	definePrimitive(base, index++, "_TempoClock_BeatsToSecs", prTempoClock_BeatsToSecs, 2, 0);
 	definePrimitive(base, index++, "_TempoClock_SecsToBeats", prTempoClock_SecsToBeats, 2, 0);
-
-	definePrimitive(base, index++, "_LinkClock_New", prClock_New<LinkClock>, 4, 0);
-	definePrimitive(base, index++, "_LinkClock_SetBeats", prClock_SetBeats<LinkClock>, 2, 0);
-	definePrimitive(base, index++, "_LinkClock_SetTempoAtBeat", prClock_SetTempoAtBeat<LinkClock>, 3, 0);
-	definePrimitive(base, index++, "_LinkClock_SetTempoAtTime", prClock_SetTempoAtTime<LinkClock>, 3, 0);
-	definePrimitive(base, index++, "_LinkClock_SetAll", prClock_SetAll<LinkClock>, 4, 0);
-	definePrimitive(base, index++, "_LinkClock_NumPeers", prLinkClock_NumPeers, 1, 0);
-	definePrimitive(base, index++, "_LinkClock_SetQuantum", prLinkClock_SetQuantum, 2, 0);
 
 	definePrimitive(base, index++, "_SystemClock_Clear", prSystemClock_Clear, 1, 0);
 	definePrimitive(base, index++, "_SystemClock_Sched", prSystemClock_Sched, 3, 0);
